refactor: split user/service_crud_test.go from service_test.go
